Revision 3. Attempt fix yay

+++ yay.sh
@@ -1,13 +1,14 @@
 #!/bin/bash
 set -e
 
-echo "Updating system..."
-sudo pacman -Syu --noconfirm
-echo "Installing go and dependencies..."
-sudo pacman -S --needed --noconfirm git base-devel go
-echo "Installing yay..."
-git clone https://aur.archlinux.org/yay.git
-cd yay && makepkg -si
-cd .. && rm -rf yay
-
-echo "Yay installed!"
+if ! command -v yay &>/dev/null; then
+  echo "Installing yay..."
+  sudo pacman -Syu --noconfirm
+  sudo pacman -S --needed --noconfirm git base-devel go
+  git clone https://aur.archlinux.org/yay.git
+  cd yay && makepkg -si
+  cd .. && rm -rf yay
+  echo "Yay installed!"
+else
+  echo "Yay already installed!"
+fi
